implement the random selection of a color/sound
the testcases are now passing as intended.

notice the game design document didn't state the random selection had to be fair.
therefore the testcase doesn't check that and merely ensures the new entry is valid.

in the interest of making the game, well.. interesting, the implementation went
overboard to ensure fairness, but returning 'blue' all the time would also have
been a valid if not predictable solution.