implement random color picking
the test cases are now passing as intended.

the game design doesn't state anything about fairness in the selection process
and I'm not going to test for that because it's not important to me. I trust
Math.random() to be fair anyway. What matters to me is that Simon only adds
valid colors to his sequence.

the goal BDD isn't to achieve 100% coverage of your codebase. you should only
test what is important to you.